requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
drf-yasg
psycopg2-binary
pillow
PyYAML
python-dotenv
opencv-python
ultralytics
pandas
numpy
gunicorn
whitenoise
celery
redis
boto3
channels
channels_redis

# TensorRT dependencies (optional)
# Note: TensorRT needs to be installed separately using NVIDIA packages
# pip install tensorrt --extra-index-url https://pypi.ngc.nvidia.com
# For development without TensorRT, these can be commented out
pycuda>=2021.1
# tensorrt>=8.0.0  # Install via NVIDIA packages

# Additional ML dependencies for TensorRT
onnx>=1.12.0
onnxruntime-gpu>=1.12.0